+++ packages/form/lib/src/extensions.dart
@@ -1,5 +1,6 @@
 library extensions;
 
 export 'extensions/form_controller_extension.dart';
+export 'extensions/form_controller_text_field_extension.dart';
 export 'extensions/form_state_field_extension.dart';
 export 'extensions/form_state_value_extension.dart';