+++ packages/form/packages/bond_form_riverpod/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.3+4
+- Bumped `bond_form` dependency to version `^0.0.6+2`.
+
+## 0.0.3+3
+- Bumped `bond_form` dependency to version `^0.0.6+1`.
+
 ## 0.0.3+2
 - Bumped `bond_form` dependency to version `^0.0.6`.